Add files via upload
Look at the main to make sure you're using the right methods to run the program. The methods that you have to use to put in your partially sudoku board are clearly labeled. in main.java you will have to use line 14 and 19 i doubled slashed them but undouble slash to do so.